update 90% function create bootcamp

+++ Client/src/components/CreateBootcamp/Semester/Semester.jsx
@@ -4,8 +4,10 @@ import Highlighter from 'react-highlight-words';
 import { Button, Input, Space, Table, Card, Divider, Tag, Row, Col } from 'antd';
 import { useDispatch, useSelector } from 'react-redux';
 import { deleteSemester, removeSubjectFromSemester } from '../../../redux/CreateBootcamp/createBootCamp';
+import { MISSING_SUBJECT_IN_SEMESTER } from '../../../util/constants/errorMessage';
+import { deleteConfirmConfig } from '../../../util/ConfirmModal/confirmConfig';
 
-const Semester = ({setIsModalOpen, subjectList, semesterIndex, setSelectedSemester}) => {
+const Semester = ({error,totalSemester,setIsModalOpen, subjectList, semesterIndex, setSelectedSemester, confirmModal}) => {
    
     const dispatch = useDispatch()
     const [searchText, setSearchText] = useState('');
@@ -151,7 +153,9 @@ const Semester = ({setIsModalOpen, subjectList, semesterIndex, setSelectedSemest
                 <Row>
                     <Col span={12}>
                     <Button type='default' danger 
-                    onClick={() => {
+                    onClick={async () => {
+                        const confirmed = await confirmModal.confirm(deleteConfirmConfig)
+                        if(confirmed)
                         dispatch(removeSubjectFromSemester({subjestIndex: data.key,semesterIndex, semesterSubjectListIndex: data.semesterSubjectListIndex}))
                     }}
                     >
@@ -203,16 +207,24 @@ const Semester = ({setIsModalOpen, subjectList, semesterIndex, setSelectedSemest
                 <Button type="primary" onClick={handleSemesterClick}>
                     Add Subject
                 </Button>
-                <Button style={{marginLeft:16}} type='default' danger 
-                    onClick={() => {
-
+                {
+                    totalSemester > 1 ?
+                    <Button style={{marginLeft:16}} type='default' danger 
+                    onClick={async () => {
+                        const confirmed = await confirmModal.confirm(deleteConfirmConfig)
+                        if(confirmed)
                         dispatch(deleteSemester(semesterIndex))
                     }}
                     >
                         Delete Semester
                 </Button>
+                :
+                <></>
+                }
+                
                 </div>
             </div>
+            {(error && subjectList.length ===0) ? <div style={{ color: "red", marginTop: 10}}>**{MISSING_SUBJECT_IN_SEMESTER}</div> : ""}
             <Divider />
             <Table columns={columns} dataSource={data} />;
         </Card>
